Create boilerplate for recommended items
