+++ src/js/api/record_item_click.tsx
@@ -1,0 +1,9 @@
+import buildWrappedGet from "../base/api"
+
+export type Arguments = {
+  itemId: number
+}
+
+export type Response = {}
+
+export const apiCall = buildWrappedGet<Arguments, Response>("/api/record_item_click")
