--- src/js/base/types.tsx
@@ -1,8 +1,0 @@
-export interface OnChangeObject {
-  value: string
-}
-
-export enum ActionType {
-  SELECT = "select",
-  DELETE = "delete",
-}
