+++ src/js/listings/recommended_items.tsx
@@ -1,0 +1,12 @@
+import * as React from "react"
+import { Row, Col } from "react-bootstrap"
+
+export default function RecommendedItems() {
+  return (
+    <Row>
+      <Col xs={12}>
+        <div />
+      </Col>
+    </Row>
+  )
+}
